+++ InventoryService.Messages/BackUpAllInventoryFailedMessage.cs
@@ -2,5 +2,11 @@
 {
     public class BackUpAllInventoryFailedMessage : BackUpAllInventoryCompletedMessage
     {
+        public BackUpAllInventoryFailedMessage(string failureMessage)
+        {
+            FailureMessage = failureMessage;
+        }
+
+        public string FailureMessage { get; }
     }
 }
