+++ InventoryService.Server/App.config
@@ -9,6 +9,8 @@
     <add key="ServerEndPoint" value="http://*:10080/" />
     <add key="EsentStoragePath" value="../" />
     <add key="ServerActorSystemName" value="InventoryService-Server" />
+    <add key="AzureStorageContainerReference" value="inventory-service-backups" />
+    <add key="AzureStorageConnectionString" value="DefaultEndpointsProtocol=https;AccountName=occ;AccountKey=occ" />
     <add key="RemoteInventoryActorAddress" value="akka.tcp://InventoryService-Server@localhost:10000/user/InventoryActor" />
   </appSettings>
   <startup>
