+++ InventoryService.ServiceDeployment/App.config
@@ -10,6 +10,8 @@
     <add key="EsentStoragePath" value="../" />
     <add key="ServerActorSystemName" value="InventoryService-Server" />
     <add key="RemoteInventoryActorAddress" value="akka.tcp://InventoryService-Server@localhost:10000/user/InventoryActor" />
+    <add key="AzureStorageContainerReference" value="inventory-service-backups" />
+    <add key="AzureStorageConnectionString" value="DefaultEndpointsProtocol=https;AccountName=occ;AccountKey=occ" />
   </appSettings>
   <connectionStrings>
   </connectionStrings>
