+++ InventoryService.Tests/app.config
@@ -9,6 +9,8 @@
     <add key="RemoteInventoryActorAddress" value="akka.tcp://InventoryService-Server@localhost:10000/user/InventoryActor" />
     <add key="ServerEndPoint" value="http://*:10086/" />
     <add key="ServerActorSystemName" value="InventoryService-Server" />
+    <add key="AzureStorageContainerReference" value="inventory-service-backups" />
+    <add key="AzureStorageConnectionString" value="DefaultEndpointsProtocol=https;AccountName=occ;AccountKey=occ" />
   </appSettings>
   <akka>
     <hocon>
